move app.css and index.css to styles folder
