+++ src/containers/App.js
@@ -1,4 +1,4 @@
-import '../App.css';
+import '../styles/App.css';
 import Routes from './Routes';
 
 const App = () => (
