+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
+import './styles/index.css';
 import { Provider } from 'react-redux';
 import App from './containers/App';
 import store from './redux/store';
